This PR modifies the `SubmitTask` and `Finish` methods of MapNode in `ExecPlan` to avoid scheduling extra thread tasks.

Performed the TPC-H Benchmark developed in PR #12537 with and without the changes.
```
TPC-H Benchmark (With Extra Thread Tasks)
-------------------------------------------------------------------
Benchmark                         Time             CPU   Iterations
-------------------------------------------------------------------
BM_Tpch_Q1/ScaleFactor:1   95035633 ns       178700 ns          100


TPC-H Benchmark (Without Extra Thread Tasks)
-------------------------------------------------------------------
Benchmark                         Time             CPU   Iterations
-------------------------------------------------------------------
BM_Tpch_Q1/ScaleFactor:1   91511754 ns       182060 ns          100

```

Also, tested with the Query Tester as proposed in PR #12586 
```
With Thread Tasks (batch size = 4096)
./query_tester tpch-1
Average       Duration: 0.106694s (+/- 0s)
Average Output  Rows/S: 37.4902rps
Average Output Bytes/S: 4573.81bps

Without Thread Tasks (batch size = 4096)
./query_tester tpch-1
Average       Duration: 0.104658s (+/- 0s)
Average Output  Rows/S: 38.2198rps
Average Output Bytes/S: 4662.82bps
```